Move otproto to otlp:internal, so can be marked as stable
