+++ exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/ResourceAdapter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.extension.otproto;
+package io.opentelemetry.exporter.otlp.internal;
 
 import io.opentelemetry.proto.resource.v1.Resource;
