+++ exporters/otlp/common/src/test/java/io/opentelemetry/exporter/otlp/internal/ResourceAdapterTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.extension.otproto;
+package io.opentelemetry.exporter.otlp.internal;
 
 import static io.opentelemetry.api.common.AttributeKey.booleanKey;
 import static io.opentelemetry.api.common.AttributeKey.doubleKey;
