--- sdk-extensions/otproto/gradle.properties
@@ -1,1 +1,0 @@
-otel.release=alpha
