+++ settings.gradle.kts
@@ -52,6 +52,7 @@ include(":exporters:jaeger-thrift")
 include(":exporters:logging")
 include(":exporters:logging-otlp")
 include(":exporters:otlp:all")
+include(":exporters:otlp:common")
 include(":exporters:otlp:metrics")
 include(":exporters:otlp:trace")
 include(":exporters:prometheus")
@@ -71,7 +72,6 @@ include(":sdk-extensions:async-processor")
 include(":sdk-extensions:autoconfigure")
 include(":sdk-extensions:aws")
 include(":sdk-extensions:logging")
-include(":sdk-extensions:otproto")
 include(":sdk-extensions:resources")
 include(":sdk-extensions:tracing-incubator")
 include(":sdk-extensions:jaeger-remote-sampler")
